+++ index.html
@@ -24,9 +24,9 @@
         
     </div>
     
-    <script type="module" src="src/editor.js" type="text/javascript"></script>
+    <script type="module" src="./src/input/editor.js" type="text/javascript"></script>
     <script type="module">
-        import { onClickRunCodeButton } from './src/editor.js';
+        import { onClickRunCodeButton } from './src/input/editor.js';
 
         document.getElementById('editor-button').onclick = function() {
             onClickRunCodeButton();
