+++ src/input/editor.js
@@ -1,18 +1,11 @@
-import { moveBunny } from "./index.js"
+import { runPythonCode} from "./pyodide.js"
 
-/* global loadPyodide */
+
 
 const ACE_LINK = "https://cdnjs.cloudflare.com/ajax/libs/ace/1.4.14/ace.js";
-var pyodide;
 var editor;
 
-initializePyodide();
 LoadAce();
-
-async function initializePyodide() {
-    pyodide = await loadPyodide();
-    pyodide.setStdin();
-}
 
 function LoadAce() {
     let script = document.createElement("script");
@@ -36,37 +29,4 @@ function initializeEditor() {
 
 export function onClickRunCodeButton() {
     runPythonCode(editor.getValue());
-}
-
-function runPythonCode(string) {
-    registerJSModules(); // poista kommentti jos poistat puputesti.py kommentit
-
-    let pythonFileStr = GetPythonFile();
-    pyodide.runPython(pythonFileStr);
-
-    pyodide.runPython(string);
-
-    let lista = pyodide.globals.get("liikelista").toJs();
-    console.log(lista);
-}
-
-function GetPythonFile() {
-    let path = "src/puputesti.py";
-    return GetFileAsText(path);
-}
-
-function GetFileAsText(filepath) {
-    let request = new XMLHttpRequest();
-    request.open('GET', filepath, false); // false = ei async
-    request.send(null);
-
-    if (request.status === 200) {
-        return request.responseText;
-    } else {
-        throw new Error(`Error fetching file: ${filepath}`);
-    }
-}
-
-function registerJSModules() { // poista kommentti kun käytetään registerJSModules
-    pyodide.registerJsModule('bunny_module', { moveBunny });
 }
